# area_back_end/web-services/docs/swagger.yaml
definitions:
  AREA_management.Area:
    properties:
      action:
        type: string
      action_name:
        type: string
      color:
        type: string
      id:
        type: integer
      name:
        type: string
      reaction:
        type: string
      reaction_name:
        type: string
      toggled:
        type: boolean
    type: object
  AREA_management.AreaCreationErrorResponse:
    properties:
      message:
        example: AREA not valid
        type: string
    type: object
  AREA_management.AreaCreationRequest:
    properties:
      action:
        example: spotify
        type: string
      action_name:
        example: new_song
        type: string
      color:
        example: '#FF0000'
        type: string
      name:
        example: agagou
        type: string
      reaction:
        example: email
        type: string
      reaction_name:
        example: send_mail
        type: string
      time_execute:
        example: '*'
        type: string
      toggled:
        type: boolean
    type: object
  AREA_management.AreaCreationSuccessResponse:
    properties:
      id:
        type: integer
      message:
        example: AREA created successfully
        type: string
    type: object
  AREA_management.DeleteAreaRequest:
    properties:
      area_id:
        type: integer
    required:
    - area_id
    type: object
  AREA_management.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /create-area:
    post:
      consumes:
      - application/json
      description: Create a new AREA with specified parameters.
      parameters:
      - description: API Key
        in: header
        name: x-api-key
        required: true
        type: string
      - description: Bearer Token
        in: header
        name: bearer
        required: true
        type: string
      - description: AREA Creation Data
        in: body
        name: area
        required: true
        schema:
          $ref: '#/definitions/AREA_management.AreaCreationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: AREA created successfully
          schema:
            $ref: '#/definitions/AREA_management.AreaCreationSuccessResponse'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/AREA_management.AreaCreationErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create a new AREA
      tags:
      - AREA
  /delete-area:
    post:
      consumes:
      - application/json
      description: Delete an existing AREA by its ID.
      parameters:
      - description: API key for access
        in: header
        name: x-api-key
        required: true
        type: string
      - description: Bearer token for user authentication
        in: header
        name: bearer
        required: true
        type: string
      - description: Request body for deleting an AREA
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/AREA_management.DeleteAreaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: AREA deleted successfully'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'error: error message'
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'message: invalid API key or Token not valid'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 'message: Error while deleting AREA please try again'
          schema:
            additionalProperties: true
            type: object
      summary: Delete an AREA
      tags:
      - AREA
  /get-area:
    post:
      consumes:
      - application/json
      description: Retrieves all the areas associated with the user identified by
        the provided token.
      parameters:
      - description: Bearer Token
        in: header
        name: bearer
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful retrieval of areas
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request data
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Invalid API key or bearer token
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all areas for a user
      tags:
      - AREA
...
  /modify-area:
    post:
      consumes:
      - application/json
      description: Modify an existing AREA with new details provided by the user.
      parameters:
      - description: Bearer token for authorization
        in: header
        name: bearer
        required: true
        type: string
      - description: API key for access control
        in: header
        name: x-api-key
        required: true
        type: string
      - description: Area information to be modified
        in: body
        name: area
        required: true
        schema:
          $ref: '#/definitions/AREA_management.Area'
      produces:
      - application/json
      responses:
        "201":
          description: AREA modified successfully
          schema:
            $ref: '#/definitions/AREA_management.Area'
        "400":
          description: Bad request when the parameters are not filled
          schema:
            $ref: '#/definitions/AREA_management.ErrorResponse'
        "401":
          description: Unauthorized if the token is not valid or missing
          schema:
            $ref: '#/definitions/AREA_management.ErrorResponse'
        "500":
          description: Internal Server Error when AREA modification fails
          schema:
            $ref: '#/definitions/AREA_management.ErrorResponse'
      summary: Modify an existing AREA
      tags:
      - AREA
...
swagger: "2.0"